requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.30
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.2
bleach==6.1.0
python-dotenv==1.0.1